Constraints files have structs and impl merged now
