+++ src/constraints/mod.rs
@@ -4,6 +4,10 @@ mod ball2;
 mod no_constraints;
 mod rectangle;
 
+pub use ball2::Ball2;
+pub use no_constraints::NoConstraints;
+pub use rectangle::Rectangle;
+
 /// A set which can be used as a constraint
 ///
 /// This trait defines an abstract function that allows to compute projections
@@ -13,25 +17,6 @@ pub trait Constraint {
     ///
     /// The given vector `x` is updated with the projection on the set
     fn project(&self, x: &mut [f64]);
-}
-
-/// The whole space, no constraints
-pub struct NoConstraints {}
-
-/// A Eucledian ball
-pub struct Ball2 {
-    centre: Option<Vec<f64>>,
-    radius: f64,
-}
-
-///
-/// A rectangle
-///
-/// A set of the form `{x : xmin <= x <= xmax}`, where `<=` is meant in the
-/// element-wise sense and either of `xmin` and `xmax` can be equal to infinity.
-pub struct Rectangle {
-    xmin: Option<Vec<f64>>,
-    xmax: Option<Vec<f64>>,
 }
 
 /* ---------------------------------------------------------------------------- */
